requirements.txt
requests
python-dotenv
stripe
pandas
dateutil
flask
fastapi
cryptography